+++ Backend/getRandom.php
@@ -1,0 +1,6 @@
+<!doctype html>
+<html lang="en">
+  <?php require 'functions.php';?>
+  <?php noCache(); ?>
+  <?php getRandomComic(); ?>
+</html>
