+++ README.md
@@ -1,5 +1,4 @@
-This website has three pages. It is written using php as the server side language.
-It has a Home, Todays Comic, and Random Comic page. The Random Comic page implements jQuery and Ajax.
-
-
-dc706_propane
+This website is for us to learn how to write in PHP and be able to work in team using Freedcamp and Github
+This website has 4 pages, home page, results page, About-us page, and contact-us
+This website has OMDB API interface to look up movies 
+My Heroku Link: https://team-alpha-web-site.herokuapp.com//?page=Menu
