+++ Webpages/404.phtml
@@ -1,0 +1,2 @@
+<p>404 - page not found.</p>
+<p> <img src="Images/Cat_not_found.JPG" alt=""> </p>
