+++ Webpages/contact-us.php
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="en">
+  <?php require 'functions.php';?>
+  <?php noCache(); ?>
+  <!-- Optional JavaScript -->
+  <!-- jQuery first, then Popper.js, then Bootstrap JS -->
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
+  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>Contact-us</title>
+  </head>
+
+  <body>
+        <?php nav_bar(); ?>
+        <?php search_bar(): ?>
+    <main>
+      <h1>Contact-us </h1>
+      <p>You can email us at <a href= "mailto:abc@example.com">Send Email</a> </p>
+      <p>or call us at (555)555-5555</p>
+    </main>
+    <footer>
+      &copy; 2019 Team-Alpha
+    </footer>
+  </body>
+</html>
