+++ index.php
@@ -1,3 +1,3 @@
 <html>
-<meta http-equiv="refresh" content="0; url=/Home.php" />
+<meta http-equiv="refresh" content="0; url=/Webpages/Home.php" />
 </html>
