+++ old/content/500.phtml
@@ -1,0 +1,1 @@
+<p> Eror 500 Forbidden page  </p>
