+++ old/content/hello-world.phtml
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<head>
+	<title>"Salutations!"</title>
+	<link rel="stylesheet" href="css/master.css">
+	<link rel="stylesheet" href="css/HomeImg.css">
+	<!-- Adam was hungry so he added in a website he made that has a lot of food on it -->
+	<style>
+		body {
+			background-color: #2D2D2D;
+		}
+
+		h1 {
+			color: #4704A1;
+			font-size: 100px;
+			font-family: Menlo, Monaco, fixed-width;
+		}
+
+		h2 {
+			color: #5215B6;
+			font-size: 80px;
+			font-family: Menlo, Monaco, fixed-width;
+		}
+
+		h3 {
+			color: #6627D7;
+			font-size: 60px;
+			font-family: Menlo, Monaco, fixed-width;
+		}
+
+		h4 {
+			color: #7A4BED;
+			font-size: 40px;
+			font-family: Menlo, Monaco, fixed-width;
+		}
+
+
+		p {
+			color: turquoise;
+			font-family: "Source Code Pro", Menlo, Monaco, fixed-width;
+		}
+	</style>
+</head>
+<body>
+	<h1>Hello World</h1>
+	<h2>Hello World</h2>
+	<h3>Hello World</h3>
+	<h4>Hello World</h4>
+	<p>Now Taylor is the one who does the pushing.</p>
+	<nav>
+		<ul>
+			<li><a href="Home.html">Home</a></li>
+			<li><a href="Menu.html">Menu</a></li>
+			<li><a href="Catering.html">Catering</a></li>
+			<li><a href="Specials.html">Specials</a></li>
+			<li><a href="Feedback.html">Feedback</a></li>
+		<!--  <li><a href="Carears.html">Carears</a></li> -->
+		</ul>
+</body>
+</html>
