+++ old/content/home.phtml
@@ -1,0 +1,72 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>Yamazaki</title>
+    <link rel="stylesheet" href="template/master.css">
+    <link rel="stylesheet" href="template/HomeImg.css">
+  </head>
+  <body>
+    <div id="innerPane">
+    <header>
+      <h1>
+        Yamazaki
+      </h1>
+      <h2>
+        Sushi & Hibachi
+      </h2>
+    </header>
+<!--    <nav>
+      <ul>
+        <li><a href="Home.html">Home</a></li>
+        <li><a href="Menu.html">Menu</a></li>
+        <li><a href="Catering.html">Catering</a></li>
+        <li><a href="Specials.html">Specials</a></li>
+        <li><a href="Feedback.html">Feedback</a></li> -->
+      <!--  <li><a href="Carears.html">Carears</a></li> --><!--
+      </ul>
+    </nav> -->
+      <h1 id="tel">
+        Tel: 706-940-0688
+      </h1>
+
+        <div class= "ImgAline">
+          <div class="boxI">
+          <img src="Images/StoreFrount.jpg" alt="" class="HomeImage">
+          </div>
+          <div class="boxT">
+            <h2>
+              We are located at:
+              <br>
+              6516 Kitten Lake Drive E4
+              <br>
+              Midland, GA 31820
+            </h2>
+              <!--Put map here-->
+            <table id="hours">
+              <caption>Store Hours</caption>
+              <tr>
+              <td>Mon:</td><td> 11:00am-9:30pm</td>
+              </tr><tr>
+              <td>Teu:</td><td> 11:00am-9:30pm</td>
+              </tr><tr>
+              <td>Wed:</td><td> 11:00am-9:30pm</td>
+              </tr><tr>
+              <td>Thr:</td><td> 11:00am-9:30pm</td>
+              </tr><tr>
+              <td>Fri:</td><td> 11:00am-10:30pm</td>
+              </tr><tr>
+              <td>Sat:</td><td> 11:00am-10:30pm</td>
+              </tr><tr>
+              <td>Sun:</td><td> 11:00am-9:30pm</td>
+              </tr>
+            </table>
+          </div>
+        </div>
+        </div>
+    <footer>
+      &copy; Yamazaki 2019
+    </footer>
+
+  </body>
+</html>
