+++ old/index.php
@@ -1,0 +1,9 @@
+<?php
+error_reporting(E_ALL);
+ini_set('display_errors', 1);
+
+require 'includes/config.php';
+require 'includes/functions.php';
+
+init();
+
