+++ old/template/master.css
@@ -1,0 +1,104 @@
+/*for whole website*/
+*{
+  color: red;
+  box-sizing: border-box;
+  /*font-family: 'courier new';
+  font-size: 1.02em;*/
+}
+
+header>h1{
+  text-align: center;
+  font-size: 4em;
+  padding-bottom: 0px;
+  margin-bottom: 0px;
+}
+header>h2{
+  padding-top: 0px;
+  margin-top: 1px;
+  text-align: center;
+  font-size: 1.8em;
+}
+nav{
+  text-align: center;
+  font-weight: bold;
+  text-decoration: none;
+  white-space:nowrap;
+  /*position: fixed;*/
+}
+nav ul{
+  list-style-type: none;
+}
+nav ul li{
+  display: inline;
+  padding-left: 5%;
+  padding-right: 5%;
+  font-size: 1.4em;
+}
+#tel {
+  text-align: center;
+}
+body{
+  background-image: url("../images/LowFire.jpg");
+  background-repeat: repeat;
+  background-color: none;
+}
+table>caption{
+  font-size: 2em;
+}
+Table#hours>caption{
+  float: right;
+}
+#innerPane{
+  margin-left: 10%;
+  margin-right: 10%;
+  padding-left: 5%;
+  padding-right: 5%;
+  padding-top: 1%;
+  border: 2px;
+  background-color: none;
+}
+
+#UnderConst{
+  text-align: center;
+  font-size: 6em;
+}
+
+.foodMenu{
+  text-align: center;
+}
+
+/*.HeadNav{
+  position: -webkit-sticky; /* Safari *//*
+  position: sticky;
+  top: 0;
+  background-color: black;
+  margin-left: 10%;
+  margin-right: 10%;
+}
+.bodyContent{
+  position: relative;
+}
+*/
+footer{
+  padding-left: 10%;
+  padding-right: 10%;
+  text-align: center;
+  padding-top: 25px;
+}
+main{
+  margin-left: 10%;
+  margin-right: 10%;
+  padding-left: 10%;
+  padding-right: 10%;
+}
+iframe{
+  margin-left: auto;
+  margin-right: auto;
+}
+#FeedbackForum{
+  display: flex;
+  align-items: center;
+  margin-left: auto;
+  margin-right: auto;
+  padding-left: 10px;
+}
