+++ old/template/specials.css
@@ -1,0 +1,17 @@
+table{
+  margin-left: auto;
+  margin-right: auto;
+  padding-top: 20px;
+}
+th{
+  font-size: 30px;
+
+}
+td.offers{
+  border-color: red;
+  border-style: dashed;
+  border-width: thin;
+}
+td{
+  text-align: center;
+}
