+++ old/template/template.php
@@ -1,0 +1,28 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <title><?php page_title(); ?> | <?php site_name(); ?></title>
+    <link href="/template/master.css" rel="stylesheet" type="text/css" />
+</head>
+<body>
+<div class="wrap">
+
+    <header>
+        <h1><?php site_name(); ?></h1>
+        <nav class="menu">
+            <?php nav_menu(); ?>
+        </nav>
+    </header>
+
+    <article>
+        <h2><?php page_title(); ?></h2>
+        <?php page_content(); ?>
+    </article>
+
+    <footer>
+        <small>&copy;<?php echo date('Y'); ?> <?php site_name(); ?>.<br><?php site_version(); ?></small>
+    </footer>
+</div>
+</body>
+</html>
