--- push.bat
@@ -1,4 +1,0 @@
-git add .
-set /P commit="Enter commit: "
-git commit -m ". %commit%"
-git push
