--- style.css
@@ -1,16 +1,0 @@
-body {
- /* padding-bottom:60px;   /* Height of the footer */ */
- background-color: #7c7c7d;
-}
-footer {
- text-align: center;
- bottom:0;
- width:100%;
- height:60px;   /* Height of the footer */
- background-color: gray;
-}
-
-.center {
-text-align: center;
-align-content: center;
-}
